gk: delete files
gk: completely remove the parser module
