+++ buildSrc/build.gradle.kts
@@ -17,26 +17,11 @@ repositories {
 val rootDir = projectDir.parentFile!!
 val parserDir = rootDir.resolve("parser")
 val genDir = parserDir.resolve("src/main/java")
-val parserPackageName = "org.aya.parser"
-val parserLibDir = genDir.resolve(parserPackageName.replace('.', '/')).absoluteFile
 
 val copyModuleInfo = tasks.register<Copy>("copyModuleInfo") {
   group = "build setup"
   from(parserDir.resolve("module-info.java"))
   into(genDir)
-}
-
-tasks.withType<AntlrTask>().configureEach antlr@{
-  outputDirectory = genDir
-  copyModuleInfo.get().dependsOn(this@antlr)
-  doFirst { parserLibDir.mkdirs() }
-  arguments.addAll(
-    listOf(
-      "-package", parserPackageName,
-      "-no-listener",
-      "-lib", "$parserLibDir",
-    ),
-  )
 }
 
 tasks.named("build").configure {
@@ -46,7 +31,6 @@ tasks.named("build").configure {
 dependencies {
   val deps = Properties()
   deps.load(rootDir.resolve("gradle/deps.properties").reader())
-  antlr("org.antlr", "antlr4", deps.getProperty("version.antlr"))
   api("org.aya-prover.upstream", "build-util", deps.getProperty("version.aya-upstream"))
 
   // The following is required for
