+++ parser/src/main/java/org/aya/parser/ij/GenericNode.java
@@ -1,3 +1,5 @@
+// Copyright (c) 2020-2022 Tesla (Yinsen) Zhang.
+// Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
 package org.aya.parser.ij;
 
 import com.intellij.AyaModified;
